Feat: made fixed mathcing, modulized generateing computer
In first, making matching function, I used 'in' for finding elements in array.
However, 'in' only find a property or a key from interable.
So, change it to 'forEach'.

Also, the play constructure changed with while and rearrange with try catch expression.

I actually forgot to commit these changes from the last commit.
So, I commited these all at once.